requirements-test.txt
django-appconf
cov-core==1.15.0
coverage<4
flake8==2.5.4
pytest==2.9.2
pytest-cov==2.2.1
pytest-django==2.9.1
wheel==0.29.0
